refactor: TicketingCommandService 인터페이스 제거 및 구체 클래스 사용
